quick-start issue 1 (upload data)
resolved issue 1 with Data Upload fails, but as a cruel joke Data
Upload is disabled, until we implement it appropriately as a feature
request.

Also made it explicit in the quick-start that we will not support
Docker for Mac, Docker for Windows due to the resource requirements of
Oracle. This may change once a Postgres database is implemented.